Check in before changing CA rules to be connection-based

+++ src/main/java/grondag/adversity/feature/volcano/lava/LavaBlockUpdate.java
@@ -5,9 +5,9 @@ import net.minecraft.util.math.BlockPos;
 public class LavaBlockUpdate
 {
     public final BlockPos pos;
-    public final float level;
+    public final int level;
     
-    public LavaBlockUpdate(BlockPos pos, float level)
+    public LavaBlockUpdate(BlockPos pos, int level)
     {
         this.pos = pos;
         this.level = level;
